requirements.txt
asgiref==3.2.3
Click==7.0
Cython==0.29.14
Django==3.0.2
dnspython==1.16.0
Flask==1.1.1
Flask-MongoAlchemy==0.7.2
Flask-PyMongo==2.3.0
gunicorn==20.0.4
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
MongoAlchemy==0.19
numpy==1.18.1
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==7.0.0
pymongo==3.10.1
pytesseract==0.3.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
tesseract==0.1.3
Werkzeug==0.16.0
